re-organizing the documentation.
Documentation is now setup to have a reference manual for SPHERLS,
SPHERLSgen, SPHERLSanal, and the python scripts. The python scripts
reference manual isn't documenting correctly as it is assuming that
the seperate scripts comprise a single combined project instead of
a number of individually functioning scripts.

There is also a seperature user guide, which should describe all
aspects of SPHERLS, while the other manuals are merely references.