Added the ability to install distributions.
This included modifying Makefile.am to tell make what data files and scripts to move for the installation. It also required modifying paths so that it worked for both the un-installed version, and the installed version of the code.

I noticed that the method for calculating the eddy viscosity had changed since the last set of reference calculations had been made, and that the reference calculations were no longer correct for the current version of the code. I have updated these to be compatible with the current version of the code.