Fixed some errors in SPHERLSanal when making 2D slices.
These errors occured when making a phi slice when the calculation was only 2D. In this case a "-" as well as the phi-velocity were written out in the 1D region (in this case a 0.0). Essentially shifting all the 1D region's contributions to the right by one column. This would only effect plotting slices in the 1D region, in the 2D case.