Added a number of reatures to the Dump class.
I have included a number of features to the dump class to allow creating rectangular arrays from variables (rather than dealing with the 1D/Mulit-D grid) and have added slicing of variables, and a number of methods to write dumps to files. I have not yet added a method to write dumps in a version that SPHERLS can read. The methods I have added are mainly for analysis purposes.