Switched form Doxygen for main user guide
Doxygen has become a pain in the ass, it no longer supports html tables correctly which is why the documentation creation was hanging. I was also having trouble simply creating a pdf but perhaps it was just too late in the day. In any case I have moved the main userguide to a plain latex file. This makes the most sense to do. In the future I may want to try making supporting reference guides from the code comments using Doxygen, a lot of the old functionality for this is still around just not active in the make script.