Added 3D interpolation to calculate light curve residuals.
I also wrote a script to find values of the distance modulus, and phase shift when comparing two light curves. However, it doesn't give reliable results and takes quite a long time. It could however be a good place to start if wanting to make such a script int he future. Likely to improve the results one would need to improve the comparison method perhaps binning the light curve that isn't being changed and using that to create a function to evaluate (using interpolation) to compute least squares from.