updated code to work with newer PETSc library 3.9.0
- added configure option "--enable-petsc-3.1-p8" which compiles code to
work with older PETSc library
- needed to specifically use mpic++ to get mpi libraries linked
properly. For some reason using g++ with older openmpi/1.8.8 seemed to
work? not sure why.
- switched from --disable-hdf to --enable-hdf as most times you don't
want that old hdf4 support
- some small differences between older library version and newer library
build in velocity of 2DNA test. Relative velocity differences as large
as 5e-13 over 10 time steps, and in the 3DNA relative differences in the
velocitis as large as 4e-11 over 10 time steps. This seems small enough
that we can say that results are still being reproduced reasonably well.